backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
numpy
tensorflow